Don't set the error message if the constructor is called with undefined
Summary:
According to the ECMAScript specification, an error "message" property
should be an empty string if the error is construced with "undefined".

Previously, Hermes assigned a value of "undefined" to error "message"
properties for the above case. This commit leaves the error "message"
property as an empty string for that case instead.